renaming urls folder so it remove python conflict
